doc: remove DNS docs for "internal service"
This is currently unused, except for the (old) builder.stage.jquery.net
pointer which I started using with the idea to do more of it, but
I can't think of another use case.

And, more importantly, even the current use case doesnt' work in practice
as it doesn't acount for gradual migration between servers.

The reason I introduced it was to avoid having to update webhooks
of many different GitHub repos, which is indeed annoying, but this
can be addressed instead by setting the webhook at the organization
level, which is what I've done for the new builder-04.stage and
wp-02.stage.ops.jquery.net servers.