miscweb: Lax "must not end with a slash" restriction
This is to fix commit efc250a37f, as otherwise Puppet fails.

I'm assuming the restriction is for the prefix case, since the
appended URL will always start with a slash. But, in the absolute
destination case, no appending takes place, and some URLs do have
to end in a slash.