P:wordpress::docs: sync jquery.org redirects
Most of these are prefix regexes, with the previous canonical URL
usually ending in a slash so the redirects did not cover those. Many
of them also used to have subpages.

For trademark and bylaws, update to the new URLs used by openjsf.org
footer, which are dynamic redirects to the latest versions (the 2019
documents have actually been obsoleted in the interim).

For contribute.jquery.org, the redirect used to be `(?i)^/CLA/`,
with as example http://contribute.jquery.org/CLA/status/. I don't
think we need lowercase /cla per se, but we definitely need /CLA.

Ref https://github.com/jquery/infrastructure/blob/66f8f5804d/modules/jquery/manifests/wp/jqueryorg.pp