wordpress: fix jquery.org search by removing "/" shadow redirect
Let WordPress handle this redirect so that it only applies when
"/" is used to target the home page, and e.g. not when it is used
for RSS permalinks like "/?p=N" or for search result pages like
"?s=X".

Similar to our nginx config, the WordPress plugin we use handles
both /foo and /foo/. I've removed the ones that used to be pages
and that have redirects set up in WordPress.

I've kept the redirects at nginx for ones that spanned several
subpages under the same prefix.