hieradata: Remove `tag_format: *semver_tag` from releases_jquery_com
Follows-up 205212dd1ec1c2d9, which accidentally added this to all sites,
including the one site where we always build from the main branch in
production as well (i.e. releases.jquery.com).

The result was that webhooks for this repo were ignored by builder-02
because it was only looking for tags, which there never are on this
repo.