P:puppet::puppetdb: allow traffic from other Puppet servers
Historically this was relying on the fact that PuppetDB is colocated on
the same servers that Puppet server itself runs on. This is still the
case, but we want to swap the active Puppet server without temporarily
losing all PuppetDB data.

For this, open a firewall rule that allows all Puppet servers to talk
to all PuppetDBs, so that a future commit can configure the Puppet
server to send reports to all available PuppetDB servers.