hieradata: [noop] Fold wordpress_version and others from envs to `common.yaml`
The description in doc/wordpress.md says it is in common.yaml,
but it wasn't actually, until now.

While at it, de-duplicate some of the other settings as well between
staging and prod, and set them centrally instead. They can always be
overridden again if they do need to differ.

Conveniently, this makes two config files redundant and thus the
whole thing a bit simpler.

Ref https://github.com/jquery/infrastructure-puppet/issues/17.