test: Fix syntax error, add `php -l` to `make lint`
Follows-up f722dd700f.

This file is only run by the periodic "smoke" job in CI, and so
didn't get noticed until a few minutes after deploying the change.
Not a big deal since the file is indeed only used by that job, and
not anywhere in prod. But.. getting that feedback earlier when there
is an obvious statically detectable mistake will be useful so let's
lint PHP files in `make lint` both locally and in CI for commit/PR.